+++ CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * Default Archetype
 * Image Gallery support
 * YouTube Video support
+* Website Cookies notification
 
 
 ### Changed